traffic_management-transport_recommendation/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
python-dotenv==1.0.1
Flask==2.3.2
networkx==3.1
osmnx==1.3.0 